% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annual_Data.r
\name{AnnualHydrometricData}
\alias{AnnualHydrometricData}
\title{Get mean annual hydrometric data for selected station(s)}
\usage{
AnnualHydrometricData(con, get_flow = TRUE, station_number)
}
\arguments{
\item{con}{An open SQLite database connection to the HYDAT database}

\item{get_flow}{Logical. If TRUE the function returns mean annual discharge,
otherwise the function returns mean annual water level}

\item{station_number}{A vector of Water Survey of Canada station identifiers}
}
\value{
A data frame (station_number, year, mean)
}
\description{
Gets the mean annual hydrometric series (flow or level) for 
   one or more stations. Flow units in m3/s; water level units in metres.
}
\seealso{
\code{\link{MonthlyHydrometricData}}

\code{\link{DailyHydrometricData}}
}

