% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_Data.r
\name{AnnualInstantaneousPeakData}
\alias{AnnualInstantaneousPeakData}
\title{Gets the annual max/min instantantaneous data for selected station(s)}
\usage{
AnnualInstantaneousPeakData(con, get_flow = TRUE, station_number)
}
\arguments{
\item{con}{An open SQLite database connection to the HYDAT database}

\item{get_flow}{Logical. If TRUE the function returns annual max/min
instantaneous flow, otherwise the function return annual max/min 
instantaneous level for each station}

\item{station_number}{A vector of Water Survey of Canada station identifiers}
}
\value{
A data frame (station_number, year, peak_code, precision_code, month,
 day, hour, minute, timezone, peak).
}
\description{
Gets the annual max/min instantaneous hydrometric series 
 (flow or level). Data are usually defined to nearest minute in local standard time
	  (see note below).
}
\note{
In some cases, the timing of the instantaneous max/min may be
 uncertain due to equipment failure (e.g., recorder failure). Event 
 timing may be censored by the hydrographer to reflect this uncertainty. This
 may lead to NA values for day or hour of occurrence. The precision_code is
 only valid for water level series and refers to the precision (mm or m) of
 the water level measurement (see \code{\link{ReportedWaterLevelPrecision}}).
}
\seealso{
\code{\link{AnnualPeakData}} for annual maximum/minimum hydrometric 
 series. See \code{\link{DataSymbols}} for information on symbols associated
 with instantaneous peak values.
}

