% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_Data.r
\name{AnnualPeakData}
\alias{AnnualPeakData}
\title{Gets the annual max/min daily data for selected station(s)}
\usage{
AnnualPeakData(con, get_flow = TRUE, station_number)
}
\arguments{
\item{con}{An open SQLite database connection to the HYDAT database}

\item{get_flow}{Logical. If TRUE the function returns the series of annual 
max/min daily flow, otherwise the function return the series of annual
max/min daily level for each station}

\item{station_number}{A vector of Water Survey of Canada station identifiers}
}
\value{
A data frame (station_number, peak, year, month day, value, flag)
}
\description{
Gets the annual maximum and minimum daily mean hydrometric 
   series (flow or level). Flow units in m3/s; water level units in metres.
}
\seealso{
\code{\link{AnnualInstantaneousPeakData}} for annual instantaneous
   records, see \code{\link{DataSymbols}} for information on symbols associated
   with peak values.
}

