% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daily_Data.r
\name{DailyHydrometricData}
\alias{DailyHydrometricData}
\title{Gets mean daily hydrometric data for selected station(s)}
\usage{
DailyHydrometricData(con, get_flow = TRUE, station_number)
}
\arguments{
\item{con}{An open SQLite database connection to the HYDAT database}

\item{get_flow}{Logical. If TRUE the function returns mean daily discharge,
otherwise the function returns mean daily water level}

\item{station_number}{A vector of Water Survey of Canada station identifiers}
}
\value{
A data frame (station_number, date, value, symbol)
}
\description{
Gets the mean daily hydrometric series (flow or level) for 
   one or more stations. Flow units in m3/s; water level units in metres.
}
\note{
Symbols associated with mean daily hydrometric series are as follows
 "A" - manual gauge
 "B" - ice affected 
 "D" - dry 
 "E" - estimated
}
\seealso{
\code{\link{AnnualHydrometricData}} for mean annual data, 
\code{\link{MonthlyHydrometricData}} for mean monthly data, 
\code{\link{DailySedimentData}} for mean daily sediment load/concentration,
\code{\link{DataSymbols}} for information on symbols associated with mean daily data, and
\code{\link{ReportedWaterLevelPrecision}} for information on water level precision
}

