% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daily_Data.r
\name{DailySedimentData}
\alias{DailySedimentData}
\title{Gets mean daily sediment data for selected station(s)}
\usage{
DailySedimentData(con, get_load = TRUE, station_number)
}
\arguments{
\item{con}{An open SQLite database connection to the HYDAT database}

\item{get_load}{Logical. If TRUE the function returns mean daily load,
otherwise the function returns mean daily concentration.}

\item{station_number}{A vector of Water Survey of Canada station identifiers}
}
\value{
A data frame (station_number, date, value).
}
\description{
Gets the mean daily suspended sediment (concentration or load)
 for one or more stations. Concentrations are measured in units of mg/L. Load 
 are measured in tonnes (mean daily conc * mean daily flow * 0.0864).
}
\note{
Symbols associated with mean daily sediment series are as follows:
<<need further work here>>
}
\seealso{
\code{\link{DailyHydrometricData}} for mean daily hydrometric data,

\code{\link{DataSymbols}} for information on symbols associated with mean daily data
}

