% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasymb.R
\docType{data}
\name{datasymb}
\alias{datasymb}
\title{Data symbols}
\format{A dataframe with 5 rows and 3 variables:
\describe{
\item{SYMBOL_ID}{Symbol ID}
\item{SYMBOL_EN}{English description of data attribute}
\item{SYMBOL_FR}{French description of data attribute}
}}
\source{
Government of Canada /url{wateroffice.ec.gc.ca}
}
\usage{
datasymb
}
\description{
Data symbols for lookup to convert letter codes to descriptions
}
\keyword{datasets}

