% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.r
\name{DataSymbols}
\alias{DataSymbols}
\title{DataSymbols}
\usage{
DataSymbols(con)
}
\arguments{
\item{con}{An open SQLite database connection}
}
\value{
A data frame  (SYMBOL_ID, DESCRIPTION)
}
\description{
Gets the description of data symbols applied to publication of daily hydrometric
 (flow,level) and suspended sediment concentrations
}

