\name{HYDAT-package}
\alias{HYDAT-package}
\alias{HYDAT }
\docType{package}
\title{
HYDAT
}
\description{
  Access to Canadian Hydrometric Data (HYDAT)
}
\details{
\tabular{ll}{
Package: \tab HYDAT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-02-01\cr
License: \tab GPL-3\cr
}
The HYDAT package requires download of the SQLite version of the Canadian hydrometric database (HYDAT) from Environment and Climate Change Canada. The database contains over 1800 active and 7000 discontinued stations. Go to the following URL for more information: \link{https://www.ec.gc.ca/rhc-wsc/default.asp?lang=En&n=9018B5EC-1}
}
\author{
Maintainer: David Hutchinson <david.hutchinson@canada.ca>
}
\references{
Environment and Climate Change Canada (2016)
}
\examples{

}
