% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat-package.R
\docType{package}
\name{hydat}
\alias{hydat}
\alias{hydat-package}
\title{hydat: clean and organize HYDAT datasets for hydro analysis}
\description{
Managing datasets downloaded from the Government of Canada's
HYDAT database can often cause trouble due to their formatting.
Why not convert them to long format and save yourself some time?
}

