% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_flooddist.R
\name{hydat_flooddist}
\alias{hydat_flooddist}
\title{Annual flood distribution}
\usage{
hydat_flooddist(x)
}
\arguments{
\item{x}{a dataframe that has been processed through hydat_load}
}
\value{
returns a dataframe of annual peak discharges and the proportion of discharges less than
}
\description{
Generates annual flood distribution by arranging annual peak flows in ascending order and calculates proportion less than
}

