% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_model_floodfreq.R
\name{hydat_model_floodfreq}
\alias{hydat_model_floodfreq}
\title{model flood recurrence intervals}
\usage{
hydat_model_floodfreq(x, logx = FALSE, logy = FALSE)
}
\arguments{
\item{x}{input dataset post-processed through `hydat_load`}

\item{logx}{logical input on whether to transform x variable with log10()}

\item{logy}{logical input on whether to transform y variable with log10()}
}
\value{
returns linear model diagnostics and model result in a list format
}
\description{
model flood recurrence intervals
}

