% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monthly_Data.r
\name{MonthlyHydrometricData}
\alias{MonthlyHydrometricData}
\title{Get mean monthly hydrometric data for selected station(s)}
\usage{
MonthlyHydrometricData(con, get_flow = TRUE, station_number)
}
\arguments{
\item{con}{An open SQLite database connection to the HYDAT database}

\item{get_flow}{Logical. If TRUE the function returns mean annual discharge,
otherwise the function returns mean annual water level}

\item{station_number}{A vector of Water Survey of Canada station identifiers}
}
\value{
A data frame (station_number, year, month, monthly_mean, monthly_total)
}
\description{
Gets the mean monthly hydrometric series (flow or levels) for 
   one or more stations. Flow units in m3/s; water level units in metres.
}
\note{
monthly_total is the total of the mean daily flow for each month that has 
   complete data whereas monthly_mean is the average of the mean daily flow for 
   each month that has complete data.
}
\seealso{
\code{\link{AnnualHydrometricData}}

\code{\link{DailyHydrometricData}}
}

