% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monthly_Data.r
\name{MonthlySediment}
\alias{MonthlySediment}
\title{Get mean monthly sediment data for selected station(s)}
\usage{
MonthlySediment(con, get_load = TRUE, station_number)
}
\arguments{
\item{con}{An open SQLite database connection to the HYDAT database}

\item{get_load}{logical; TRUE to get bedload, FALSE to get suspended load}

\item{station_number}{a vector of Water Survey of Canada station identifiers}
}
\description{
Gets the mean monthly sediment (suspended or bedload) for one or 
   more stations within the Water Survey of Canada hydrometric database (HYDAT).
   Suspended load measured in units of mg/l. Bedload measured in units of tonnes.
}

