% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sediment_Samples.r
\name{ParticleSizeDistribution}
\alias{ParticleSizeDistribution}
\title{ParticleSizeDistribution}
\usage{
ParticleSizeDistribution(con, stationID)
}
\arguments{
\item{con}{an open SQLite database connection}

\item{stationID}{A vector of Water Survey of Canada station identifiers}
}
\value{
A data frame
}
\description{
Gets the particle size distribution (in millimetres) for sediment samples collected at various
 station locations
}

