% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.r
\name{PrecisionCodes}
\alias{PrecisionCodes}
\title{PrecisionCodes}
\usage{
PrecisionCodes(con)
}
\arguments{
\item{con}{An open SQLite database connection}
}
\value{
A data frame (PRECISION_CODE, DESCRIPTION)
}
\description{
Gets the description of precision codes of water levels.
}

