% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Real_Time_Data.r
\name{RealTimeData}
\alias{RealTimeData}
\title{Gets current available data for hydrometric station}
\usage{
RealTimeData(base_url = "http://dd.weather.gc.ca/hydrometric", prov_terr_loc,
  station_number)
}
\arguments{
\item{base_url}{Base URL to access the data mart}

\item{prov_terr_loc}{Two-digit provincial or territorial abbreviation in which 
the station is located
 (e.g., "BC")}

\item{station_number}{A seven-digit station identifier (e.g., "08MF005")}
}
\value{
A data frame (station_number, date_time, hg, hg_grade, hg_sym, hg_code,
 qr, qr_grade, qr_sym, qr_code)
}
\description{
Gets near real-time (within 6 hours) hydrometric data for the 
 past 30 days for a particular hydrometric station operated by the National 
 Hydrological Service of Canada, Water Survey Division.
}
\note{
Returned date/times are in Universal Time Coordinated (UTC).
}
\examples{
# get realtime data for Fraser River at Hope (08MF005)
data <- RealTimeData(prov_terr_loc="BC", station_number="08MF005")
plot(
  data$date_time, data$qr,
 type="l", col="blue",
  xlab="DateTime [UTC]", ylab="Discharge")
}
\seealso{
\code{\link{RealTimeNetwork}}
}

