% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Real_Time_Data.r
\name{RealTimeNetwork}
\alias{RealTimeNetwork}
\title{Get list of available hydrometric stations with near real-time data}
\usage{
RealTimeNetwork(base_url = "http://dd.weather.gc.ca/hydrometric/doc",
  master_file = "hydrometric_StationList.csv")
}
\arguments{
\item{base_url}{Base URL to access data mart documentation}

\item{master_file}{Filename containing the list of available real-time stations}
}
\value{
A data frame (station_number, name, latitude, longitude, prov_terr_loc, timezone)
}
\description{
Gets the current list of hydrometric gauges available in 
 near real-time (within 6 hours) from the National Hydrological Service, 
 Environment and Climate Change Canada.
}
\examples{
network <- RealTimeNetwork()
# show top 50 stations 
head(network, 50)
}
\seealso{
\code{\link{RealTimeData}}
}

