% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.r
\name{ReportedWaterLevelPrecision}
\alias{ReportedWaterLevelPrecision}
\title{ReportedWaterLevelPrecision}
\usage{
ReportedWaterLevelPrecision(con, stationID)
}
\arguments{
\item{con}{an open SQLite database connection}

\item{stationID}{A vector of Water Survey of Canada station identifiers}
}
\value{
A data frame
}
\description{
Returns stated precision of water level (cm or mm) for each year of
 published record for a given hydrometric station.
}

