% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sediment_Samples.r
\name{SedimentSamples}
\alias{SedimentSamples}
\title{Get miscellaneous sediment data}
\usage{
SedimentSamples(con, stationID)
}
\arguments{
\item{con}{an open SQLite database connection}

\item{stationID}{a vector of Water Survey of Canada station identifiers}
}
\value{
A data frame
}
\description{
Gets sediment samples collected by Water Survey of Canada staff at
various station locations. Sediment samples are categorized based on the sampling
method (Environment Canada, 1988).\cr\cr
Depth-Integrating (DI) Method: Suspended sediment load estimated from 'n'
equal-discharge segments in which sediment is sampled vertically through the water
column at a uniform rate.\cr\cr
Point-Integating (PI) Method: Suspended sediment load estimated from graphical methods.\cr
Bed Material (BM): Samples of bed material often taken in same sampling verticals as
suspended sediment or at waters edge. The total   

obtained by either (1) Depth-Integrating
Method or (2) Single Suspended
}
\references{
Environment Canada, 1988. Miscellaneous Sediment Data 1966-1983, Inland 
 Waters Directorate, Water Resources Branch, Water Survey of Canada, Ottawa, Canada. 248 p.
}

