% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Station_Metadata.r
\name{StationDataCollection}
\alias{StationDataCollection}
\title{Gets data collection operations for selected hydrometric station(s)}
\usage{
StationDataCollection(con, station_number)
}
\arguments{
\item{con}{An open SQLite database connection to the HYDAT database}

\item{station_number}{A vector of Water Survey of Canada station identifiers}
}
\value{
A data frame (station_number, year_from, year_to, data_type, 
measurement, operation)
}
\description{
Gets information on the station data collection schedules for 
 one or more station(s). This includes observation (recorder/manual) and 
 operation (seasonal/continuous) schedules for each variable recorded at 
 each station.
}

