% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Station_Metadata.r
\name{StationDataTypes}
\alias{StationDataTypes}
\title{Gets list of variables available for hydrometric station(s)}
\usage{
StationDataTypes(con, station_number)
}
\arguments{
\item{con}{An open SQLite database connection to the HYDAT database}

\item{station_number}{A vector of Water Survey of Canada station identifiers}
}
\value{
A data frame (station_number, data_type, sed_data_type, year_from,
 year_to, record_length, data_type_name, sed_data_type_name)
}
\description{
Gets information on the variables (water level, flow, and 
 sediment) collected at a station.
}

