% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Station_Metadata.r
\name{StationMetadata}
\alias{StationMetadata}
\title{Gets relevant metadata for hydrometric station(s)}
\usage{
StationMetadata(con, station_number)
}
\arguments{
\item{con}{An open SQLite database connection to the HYDAT database}

\item{station_number}{A vector of Water Survey of Canada station identifiers}
}
\value{
A data frame (station_number, station_name, prov_ter_loc, 
hydrometric_status, sediment_status, latitude, longitude, drainage_area_gross,
drainage_area_effective, rhbn, real_time, regional_office, contributor, 
operator, datum)
}
\description{
Gets relevant station metadata such as location, current
 operating status, real-time, and whether station is part of the Reference
 Hydrometric Basin Network (RHBN).
}

