% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Station_Metadata.r
\name{StationRegulation}
\alias{StationRegulation}
\title{Gets regulation history for hydrometric station(s)}
\usage{
StationRegulation(con, station_number)
}
\arguments{
\item{con}{An open SQLite database connection to the HYDAT database}

\item{station_number}{A vector of Water Survey of Canada station identifiers}
}
\value{
A data frame (station_number, year_from, year_to, regulated) which
  outlines the period of regulation. If unregulated, then year_from will be 
  an NA value.
}
\description{
Extracts information on flow regulation history affecting a 
 hydrometric station.
}
\details{
A hydrometric station is deemed regulated if the natural flow 
 regime is modified more than X% (Need to speak to Lynne!)
}

