% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Station_Metadata.r
\name{StationRemarks}
\alias{StationRemarks}
\title{Gets ancillary remarks regarding hydrometric station data}
\usage{
StationRemarks(con, station_number)
}
\arguments{
\item{con}{An open SQLite database connection to the HYDAT database}

\item{station_number}{A vector of Water Survey of Canada station identifiers}
}
\description{
Provides further clarification on data stored in HYDAT database. 
 Remarks are somewhat variable, but can provide further clarification on 
 hydrometric series.
}

