% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\docType{data}
\name{stations}
\alias{stations}
\title{Stations}
\format{A dataframe with 7791 rows and 15 variables:
\describe{
\item{STATION_NUMBER}{Station code according to database}
\item{STATION_NAME}{Station name}
\item{PROV_TERR_STATE_LOC}{Province, state, territory location}
\item{REGIONAL_OFFICE_ID}{Regional office ID}
\item{HYD_STATUS}{Status code of hydrological data}
\item{SED_STATUS}{Status code of sediment data}
\item{LATITUDE}{Latitude}
\item{LONGITUDE}{Longitude}
\item{DRAINAGE_AREA_GROSS}{Gross drainage area in km^2}
\item{DRAINAGE_AREA_EFFECT}{Effective drainage area in km^2}
\item{RHBN}{RHBN Y/N}
\item{REAL_TIME}{Real time Y/N}
\item{CONTRIBUTOR_ID}{Contributor ID}
\item{OPERATOR_ID}{Operator ID number}
\item{DATUM_ID}{Datum ID number}
}}
\source{
Government of Canada /url{wateroffice.ec.gc.ca}
}
\usage{
stations
}
\description{
Station names, codes, and respective metadata
}
\keyword{datasets}

