% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Version.r
\name{Version}
\alias{Version}
\title{Gets current version of the HYDAT database}
\usage{
Version(con)
}
\arguments{
\item{con}{An open SQLite database connection to the HYDAT database}
}
\value{
A data frame (version, date)
}
\description{
Determines the version and publication date 
 of the HYDAT database
}

