% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydat_load.R
\name{hydat_load}
\alias{hydat_load}
\title{Clean hydat data}
\usage{
hydat_load(path, return = "discharge")
}
\arguments{
\item{path}{the input path of the .csv file}

\item{return}{select output for either "discharge" or stream "level"}
}
\value{
returns discharge or stage dataframe in long format
}
\description{
Takes in a .csv file from the hydat database and converts it into 'long' format and splits into two user-defined objects for discharge and stage
}

