% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{hydat_load}
\alias{hydat_load}
\title{Clean hydat data}
\usage{
hydat_load(path, discharge_obj = "discharge", levels_obj = "levels",
  monthly_meansQ = "mo_meanQ", monthly_meansLvl = "mo_meanLvl")
}
\arguments{
\item{path}{the input path of the .csv file}

\item{discharge_obj}{names of the object outputted for discharge}
}
\value{
returns two user-defined objects for discharge and stage dataframes
}
\description{
Takes in a .csv file from the hydat database and converts it into 'long' format and splits into two user-defined objects for discharge and stage
}

